issue with crc16 fixed (allow files larger than 256 bytes)

+++ libraries/crc/crc16.c
@@ -35,7 +35,7 @@
 
 /*---------------------------------------------------------------------------*/
 /* CRC-16-IBM, CRC-16-ANSI */
-uint16_t crc16(const uint8_t* data, uint8_t num_bytes, uint16_t init_value)
+uint16_t crc16(const uint8_t* data, uint16_t num_bytes, uint16_t init_value)
 {
   uint16_t crc = init_value;
   while(num_bytes) {
@@ -47,7 +47,7 @@ uint16_t crc16(const uint8_t* data, uint8_t num_bytes, uint16_t init_value)
       } else {
         crc >>= 1;
       }
-      ch >>= 1; 
+      ch >>= 1;
       bit--;
     }
     data++;
