+++ libraries/dpp_lib.h
@@ -72,7 +72,7 @@ typedef struct list_meta {
 
 /* --- function prototypes --- */
 
-uint16_t crc16(const uint8_t* data, uint8_t num_bytes, uint16_t init_value);
+uint16_t crc16(const uint8_t* data, uint16_t num_bytes, uint16_t init_value);
 uint32_t crc32(const uint8_t* data, uint32_t num_bytes, uint32_t seed);
 
 uint32_t hexstr_to_uint32(const char* str);
